requirements.txt
azure-functions==1.22.1
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
